requirements.txt
django
djangorestframework
django-rest-swagger
drf-nested-routers
django-rest-framework-nested
django-cors-headers
channels